requirements.txt
streamlit
chromadb
litellm
python-dotenv
tenacity
pydantic
pysqlite3-binary